# Use better appropriate 5 byte hash function