@klauspost 2022-09-19
Fails on `v4` as well 🤷🏼