@greatroar 2022-09-25
The failure is TestWriterLegacy, which is known.

Do you have a figure as to how much decompression speed suffers from the shorter matches?