@klauspost 2022-09-25
Here are some mixed input types:

CSV: [nyc-taxi-data-10M.csv](https://files.klauspost.com/compress/nyc-taxi-data-10M.csv.zst) 

2044.47 MB/s  -> 2177.17 MB/s (6 bytes better decompression speed) 
1,075,178,803 bytes  -> 1,066,961,737 bytes (6 bytes better)

XML: [enwik9](https://files.klauspost.com/compress/enwik9.zst) (which really likes short matches)

1821.98 MB/s  -> 1998.78 MB/s (6 bytes better)
454,825,169  bytes  -> 469,004,632 (5 bytes better)

DB: [consensus.db.10gb](https://files.klauspost.com/compress/consensus.db.10gb.zst)

2231.56 -> 2254.25 MB/s (6 bytes better)
5,064,885,081 -> 5,064,926,177 bytes (5 bytes better)

JSON: [github-june-2days-2019.json](https://files.klauspost.com/compress/github-june-2days-2019.json.zst):

2640.45 -> 2709.13 MB/s (6 bytes better)
1,278,906,810 -> 1,274,297,625 (6 bytes better)

6 byte hashes are consistently faster to decompress and sometimes worse, sometimes better in terms of compression.

Compression speed is roughly equivalent. IMO there are plenty of "modes" left that can offer better compression.