@greatroar 2022-09-25
Convincing. I've opened a revert at #197 as an alternative.